TARS: extract bwa scoring primitives from the overhang gate
